time is NOT a float, it must be cast to one
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v2_0branch/bzflag@12305 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/game/PlayerInfo.cxx
@@ -386,7 +386,7 @@ void PlayerInfo::setAutoPilot(bool _autopilot) {
 bool PlayerInfo::isTooMuchIdling(float kickThresh) {
   bool idling = false;
   if ((state > PlayerInLimbo) && (team != ObserverTeam)) {
-    const float idletime = (now - lastupdate);
+    const float idletime = (float)(now - lastupdate);
     if (idletime > kickThresh) {
       idling = true;
     }
